Add files via upload
This is a small project that generates your name by creating individual letters using '*' symbol in vertical fashion. just enter your name and you will get the response. It consists of  individual files from 1-26 that is the code for creating each individual alphabet. took a little bit of time but is an interesting one. It is made in java jdk 1.8 and is inspired from my first year of engineering(Computer Science)