Add files via upload
This folder was large in size so uploaded in .rar format. This is an inventory management system made in java jdk1.5, this keeps the status of the present quantity, price, company name, etc of each product that is present in a particular store . for keeping the details it uses microsoft access database.  To run it use java jdk 1.5 instead of 1.8 or java 10 cause it will show errors in netbeans